+++ bidscoin/cli/__init__.py
@@ -1,1 +1,1 @@
-# Imports from the standard library only (as these are imported during the cli/manpage build process)
+"""Argument parsers for generating BIDScoin manpages and Command Line Interfaces (CLI). NB: Use imports from the standard library only (as these are imported during the cli/manpage build process)"""