+++ bidscoin/plugins/__init__.py
@@ -1,0 +1,1 @@
+"""Pre-installed plugins"""