+++ bidscoin/schema/__init__.py
@@ -4,5 +4,6 @@ The original files are released with a Creative Commons Attribution 4.0 Internat
 
 For more information see:
 https://github.com/bids-standard/bids-specification
+https://github.com/bids-standard/bids-specification/tree/master/src/schema
 https://github.com/bids-standard/bids-specification/blob/master/LICENSE
 """