+++ bidscoin/utilities/__init__.py
@@ -1,0 +1,1 @@
+"""A package of useful helper tools"""